lm_evaluation/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.13.1
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
distro==1.9.0
filelock==3.16.1
fonttools==4.55.4
frozenlist==1.4.1
fsspec==2024.9.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==0.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.136
llamaapi==0.1.36
matplotlib==3.10.0
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
openai==1.52.0
orjson==3.10.7
pandas==2.2.3
pillow==11.1.0
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==14.2
wheel==0.44.0
xxhash==3.5.0
yarl==1.15.4
